# openapi.yaml
openapi: 3.0.3
info:
  title: ADADProject API
  version: "1.0.0"
  description: |
    REST API for ADADProject (events and users). This document describes the endpoints implemented in `routes/events.js` and `routes/users.js`.
servers:
  - url: http://localhost:3000
paths:
...
  /users:
    get:
      summary: List users (paginated)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
    post:
      summary: Create one or multiple users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/UserCreate'
                - type: array
                  items:
                    $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Insert result
          content:
            application/json:
              schema:
                type: object
  /users/{id}:
    get:
      summary: Get user by id (includes user's top 3 events)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User with top3 events
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithTop3'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted flag
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{id}/review/{event_id}:
    post:
      summary: Add or update an event rating by a user
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: User ID
        - in: path
          name: event_id
          required: true
          schema:
            type: string
          description: Event ID (ObjectId string)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: integer
                  minimum: 0
                  maximum: 5
                ratedAt:
                  type: string
                  format: date-time
                  description: Optional. Defaults to current time if not provided
              required:
                - rating
      responses:
        '201':
          description: Review added
          content:
            application/json:
              schema:
                type: object
                properties:
                  added:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
        '200':
          description: Review updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
...
  /users/active/{year}:
    get:
      summary: Active users in a year
      parameters:
        - in: path
          name: year
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Users who submitted reviews in the given year
          content:
            application/json:
              schema:
                type: object
                properties:
                  year:
                    type: integer
                  activeUserCount:
                    type: integer
                  activeUsers:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
components:
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Event:
      type: object
      properties:
        _id:
          type: string
        changeDate:
          type: string
          description: original changeDate used by the project dataset
        establishmentID:
          type: string
        establishmentName:
          type: string
        address:
          type: string
        zipCode:
          type: string
        county:
          type: string
      example:
        _id: "650a3c..."
        changeDate: "2023-05-01T00:00:00.000Z"
        establishmentID: "E123"
        establishmentName: "Sample Place"
        address: "123 Main St"
        zipCode: "12345"
        county: "SomeCounty"
...
    EventRating:
      type: object
      properties:
        eventId:
          type: string
          description: ObjectId of the event
        rating:
          type: integer
          minimum: 0
          maximum: 5
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
        date:
          type: string
          format: date-time
      example:
        eventId: "650a3c..."
        rating: 5
        timestamp: 1699401600000
        date: "2025-11-08T12:00:00.000Z"
...
    User:
      type: object
      properties:
        _id:
          type: integer
        name:
          type: string
        gender:
          type: string
          enum: [M, F]
        age:
          type: integer
        occupation:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventRating'
      example:
        _id: 1234
        name: "Alice"
        gender: "F"
        age: 30
        occupation: "Engineer"
        events: []
    UserCreate:
      type: object
      required: [name, gender, age, occupation]
      properties:
        name:
          type: string
        gender:
          type: string
          enum: [M, F]
        age:
          type: integer
        occupation:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventRating'
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        gender:
          type: string
          enum: [M, F]
        age:
          type: integer
        occupation:
          type: string
        events:
          type: array
          items:
            $ref: '#/components/schemas/EventRating'
    UserWithTop3:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        bestRatedEvents:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Event'
              - type: object
                properties:
                  rating:
                    type: integer
                    minimum: 0
                    maximum: 5
                  timestamp:
                    type: integer
                    format: int64
                  date:
                    type: string
                    format: date-time
...
    PaginatedUsers:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/User'